+++ src/components/v5/common/CompletedAction/partials/Blocks/index.ts
@@ -1,0 +1,1 @@
+export * from './Blocks';
