+++ src/components/v5/common/CompletedAction/partials/EditColonyDetails/EditColonyDetails.tsx
@@ -8,11 +8,13 @@ import { formatText } from '~utils/intl';
 import UserPopover from '~v5/shared/UserPopover';
 
 import { ICON_SIZE } from '../../consts';
-import { ActionDataGrid, ActionSubtitle, ActionTitle } from '../Blocks/Blocks';
-import ActionTypeRow from '../rows/ActionType';
-import CreatedInRow from '../rows/CreatedInRow';
-import DecisionMethodRow from '../rows/DecisionMethod';
-import DescriptionRow from '../rows/Description';
+import { ActionDataGrid, ActionSubtitle, ActionTitle } from '../Blocks';
+import {
+  ActionTypeRow,
+  CreatedInRow,
+  DecisionMethodRow,
+  DescriptionRow,
+} from '../rows';
 
 const displayName = 'v5.common.CompletedAction.partials.EditColonyDetails';
