+++ src/components/v5/common/CompletedAction/partials/rows/index.ts
@@ -1,0 +1,6 @@
+export { default as ActionTypeRow } from './ActionType';
+export { default as AmountRow } from './Amount';
+export { default as CreatedInRow } from './CreatedIn';
+export { default as DecisionMethodRow } from './DecisionMethod';
+export { default as DescriptionRow } from './Description';
+export { default as TeamFromRow } from './TeamFrom';
